requirements.txt
asgiref==3.2.10
blis==0.4.1
bugsnag==3.6.0
cached-property==1.5.1
catalogue==1.0.0
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.16.0
cymem==2.0.3
dj-database-url==0.5.0
Django==2.2.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.0.2
django-cors-middleware==1.3.1
django-environ==0.4.5
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.8
gunicorn==19.9.0
idna==2.10
mccabe==0.6.1
mock==3.0.5
more-itertools==8.4.0
murmurhash==1.0.2
numpy==1.19.0
ordered-set==3.1.1
ordered-set-stubs==0.1.3
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.2
profanity-filter==1.3.3
pycodestyle==2.5.0
pydantic==1.6.1
pyflakes==2.1.1
python-http-client==3.1.0
pytz==2019.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.15.100
sendgrid==6.0.5
six==1.12.0
spacy==2.3.2
sqlparse==0.3.0
srsly==1.0.2
thinc==7.4.1
tomlkit==0.5.11
tqdm==4.48.0
urllib3==1.25.3
wasabi==0.7.1
WebOb==1.8.5
whitenoise==4.1.2
